warnings: reach -Werror cleanliness across headers, tests, vessel
Polyfill std::add_sat/sub_sat/mul_sat (C++26 P0543) in Saturate.h —
libstdc++ 16 rawhide doesn't ship them yet.  All 11 call sites
rewired to crucible::sat::*.  Verified against P0543 spec for all
signed/unsigned edge cases (INT_MIN-INT_MIN, INT_MIN*-1, etc.).

Strong IDs/hashes now noexcept end-to-end so the implicit default
ctors of containing structs propagate nothrow.  Replace memset-on-
non-trivial-NSDMI types with placement-new or std::uninitialized_
value_construct_n / fill_n / array{} so -Wclass-memaccess stays
silent.  Remove five C-only warning flags that emitted "valid for
C/ObjC but not for C++" on every TU.  Drop -Wswitch-enum — the
Op/CKernelId enums have 60+ values and explicit default: is already
required via -Werror=switch + -Wswitch-default.

Enable CRUCIBLE_WERROR in the default preset.  Clean build, 25/25
tests pass.

+++ test/test_harness.h
@@ -22,7 +22,7 @@ inline void elevate_priority() {
     setpriority(PRIO_PROCESS, 0, -10);
     cpu_set_t cpuset;
     CPU_ZERO(&cpuset);
-    CPU_SET(sched_getcpu(), &cpuset);
+    CPU_SET(static_cast<size_t>(sched_getcpu()), &cpuset);
     sched_setaffinity(0, sizeof(cpuset), &cpuset);
 #endif
 }
